refactor(annotation): холст указаний переезжает в ui/
Указания рисуют теперь двое: снимок шага сборки в тех-карте и вложение задачи. Адаптер над
поверхностью был один и жил в `managers/tech-card`. Импортировать его в `managers/tasks`
значило бы связать два домена через голову общего слоя: правка тех-карты меняла бы поведение
задач молча, а обратная зависимость появилась бы при первой же попытке навести порядок.

ПЕРЕЕХАЛО ДВА КУСКА, оба доменно-нейтральные:

`ui/components/annotation/canvas.tsx` — перевод «значение владельца ↔ вью-модель
поверхности», сборка surface-объекта, инлайн-кадр с кнопкой зума и полноэкранный диалог.
Тип значения параметризован: `AnnotationValue` — та же структура, что лежит в JSON-колонке и
приезжает по проводу, но вид и цвет объявлены строками. Домен сужает их своей схемой (у
тех-карты zod-энумы из того же реестра видов), холст обязан пережить незнакомый ключ — реестр
разрешит его в пин, и это правило старше переезда.

`ui/components/annotation/wire.ts` — словари «вид/цвет ↔ константа провода». Жили в схеме
тех-карты, пока владелец был один; вторая копия означала бы вид, который приезжает на снимок
узла и не приезжает на вложение. Схема реэкспортирует их, поэтому ни один импорт в тех-карте
не тронут.

Сборка surface-объекта вынута в `useAnnotationSurface` — единственная правка сверх
перемещения строк, и она нужна по делу: вложение задачи открывается сразу во весь экран по
клику на плитку, инлайн-кадр под плиткой был бы лишней картинкой, а два адаптера ради этого
разошлись бы в правилах правки. `AnnotationCanvas` теперь зовёт тот же хук и ведёт себя
ровно как раньше.

В тех-карте осталась тонкая обёртка со своими типами и пикером деталей кроя.

НЕИЗМЕННОСТЬ ДОКАЗАНА ЗАМЕРОМ, а не заявлена: стенд рисует полосу из двух холстов (правимый
и замороженный) с пином, меркой, легендой и пикером деталей — до и после переезда, тремя
состояниями (полоса в покое, открытый редактор после клика по фигуре, полноэкранный вид).
Снимки совпадают побайтово во всех трёх.
